Moved metrics listeners to riptide-failsafe
Fixes #429
