+++ riptide-failsafe/pom.xml
@@ -30,6 +30,11 @@
             <version>${failsafe.version}</version>
         </dependency>
         <dependency>
+            <groupId>org.zalando</groupId>
+            <artifactId>riptide-metrics</artifactId>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-test</artifactId>
         </dependency>
