+++ riptide-failsafe/src/main/java/org/zalando/riptide/failsafe/metrics/package-info.java
@@ -1,4 +1,4 @@
 @ParametersAreNonnullByDefault
-package org.zalando.riptide.spring.metrics;
+package org.zalando.riptide.failsafe.metrics;
 
 import javax.annotation.ParametersAreNonnullByDefault;
