+++ riptide-spring-boot-starter/src/main/java/org/zalando/riptide/spring/DefaultRiptideRegistrar.java
@@ -35,8 +35,8 @@ import org.zalando.riptide.httpclient.RestAsyncClientHttpRequestFactory;
 import org.zalando.riptide.metrics.MetricsPlugin;
 import org.zalando.riptide.spring.RiptideProperties.Client;
 import org.zalando.riptide.spring.RiptideProperties.Client.Keystore;
-import org.zalando.riptide.spring.metrics.MetricsCircuitBreakerListener;
-import org.zalando.riptide.spring.metrics.MetricsRetryListener;
+import org.zalando.riptide.failsafe.metrics.MetricsCircuitBreakerListener;
+import org.zalando.riptide.failsafe.metrics.MetricsRetryListener;
 import org.zalando.riptide.stream.Streams;
 import org.zalando.riptide.timeout.TimeoutPlugin;
 import org.zalando.stups.oauth2.httpcomponents.AccessTokensRequestInterceptor;
