+++ riptide-spring-boot-starter/src/test/java/org/zalando/riptide/spring/ManualConfiguration.java
@@ -42,8 +42,8 @@ import org.zalando.riptide.faults.TransientFaultPlugin;
 import org.zalando.riptide.httpclient.RestAsyncClientHttpRequestFactory;
 import org.zalando.riptide.metrics.MetricsPlugin;
 import org.zalando.riptide.spring.PluginTest.CustomPlugin;
-import org.zalando.riptide.spring.metrics.MetricsCircuitBreakerListener;
-import org.zalando.riptide.spring.metrics.MetricsRetryListener;
+import org.zalando.riptide.failsafe.metrics.MetricsCircuitBreakerListener;
+import org.zalando.riptide.failsafe.metrics.MetricsRetryListener;
 import org.zalando.riptide.stream.Streams;
 import org.zalando.riptide.timeout.TimeoutPlugin;
 import org.zalando.stups.oauth2.httpcomponents.AccessTokensRequestInterceptor;
